refactor: dto 자료형 변경

+++ src/main/java/com/github/bluekey/dto/AdminArtistProfileListDto.java
@@ -13,13 +13,17 @@ public class AdminArtistProfileListDto {
     private final List<ArtistProfileListDto> artist;
 
     @Schema(description = "매출액", example = "300")
-    private final int revenue;
+    private final long revenue;
+
     @Schema(description = "순수익", example = "1234")
-    private final int netIncome;
+    private final long netIncome;
+
     @Schema(description = "정산액", example = "1234124")
-    private final int settlementAmount;
+    private final long settlementAmount;
+
     @Schema(description = "트랙명", example = "love")
     private final String representativeTrack;
+
     @Schema(description = "요율", example = "2.5")
-    private final int monthlyIncreaseRate;
+    private final double monthlyIncreaseRate;
 }
